# muralia-backend/openapi/muralia-api.yaml
openapi: 3.0.3
info:
  title: Muralia API
  description: Backend API for managing customer authentication and image uploads
  version: 1.0.0
  contact:
    name: Muralia API Support

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.muralia.com
    description: Production server

tags:
  - name: Authentication
    description: Customer authentication endpoints
  - name: Images
    description: Image upload and retrieval endpoints

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new customer
      description: Creates a new customer account
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Customer successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Customer already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login customer
      description: Authenticates a customer and returns a JWT token
      operationId: loginCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current customer profile
      description: Returns the profile information of the authenticated customer
      operationId: getCurrentCustomer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Customer profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout customer
      description: Logs out the current customer (optional - depends on token invalidation strategy)
      operationId: logoutCustomer
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Successfully logged out
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/images:
    post:
      tags:
        - Images
      summary: Upload a new image
      description: Uploads a new image for the authenticated customer
      operationId: uploadImage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
                title:
                  type: string
                  description: Optional title for the image
                  maxLength: 255
                description:
                  type: string
                  description: Optional description for the image
                  maxLength: 1000
      responses:
        '201':
          description: Image successfully uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '400':
          description: Invalid file or request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Images
      summary: Get latest images
      description: Retrieves the latest N images uploaded to the system
      operationId: getLatestImages
      parameters:
        - name: limit
          in: query
          description: Maximum number of images to return
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of images to skip for pagination
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of latest images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/images/{imageId}:
    get:
      tags:
        - Images
      summary: Get image by ID
      description: Retrieves a specific image by its ID
      operationId: getImageById
      parameters:
        - name: imageId
          in: path
          description: Image ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Images
      summary: Delete an image
      description: Deletes an image (only the owner can delete their images)
      operationId: deleteImage
      security:
        - bearerAuth: []
      parameters:
        - name: imageId
          in: path
          description: Image ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Image successfully deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - not the image owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/images/{imageId}/file:
    get:
      tags:
        - Images
      summary: Get image file
      description: Returns the actual image file bytes from the database
      operationId: getImageFile
      parameters:
        - name: imageId
          in: path
          description: Image ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        password:
          type: string
          format: password
          minLength: 8
          example: SecurePass123!
        firstName:
          type: string
          maxLength: 100
          example: John
        lastName:
          type: string
          maxLength: 100
          example: Doe

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: SecurePass123!

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        customer:
          $ref: '#/components/schemas/Customer'

    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: johndoe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z

    Image:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        url:
          type: string
          format: uri
          description: URL to access the image
          example: https://storage.muralia.com/images/abc123.jpg
        thumbnailUrl:
          type: string
          format: uri
          description: URL to access the thumbnail
          example: https://storage.muralia.com/thumbnails/abc123.jpg
        title:
          type: string
          example: Beautiful Sunset
        description:
          type: string
          example: A stunning sunset captured at the beach
        fileName:
          type: string
          example: sunset_2024.jpg
        fileSize:
          type: integer
          format: int64
          description: File size in bytes
          example: 2048576
        mimeType:
          type: string
          example: image/jpeg
        width:
          type: integer
          example: 1920
        height:
          type: integer
          example: 1080
        customerId:
          type: integer
          format: int64
          example: 1
        customerUsername:
          type: string
          example: johndoe
        uploadedAt:
          type: string
          format: date-time
          example: 2024-01-15T14:30:00Z

    ImageListResponse:
      type: object
      properties:
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        total:
          type: integer
          description: Total number of images available
          example: 150
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T14:30:00Z
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid input data
        path:
          type: string
          example: /api/auth/register